remove kotlin core dep on aws-api
